Refactor CLI helpers into separate modules
