+++ src/autoclean/cli.py
@@ -45,491 +45,9 @@ from autoclean.utils.task_discovery import (
 )
 from autoclean.utils.user_config import user_config
 from autoclean.utils.console import get_console
+from autoclean.cli_logging import log_cli_execution, strip_wrapping_quotes
+from autoclean.cli_help import (attach_rich_help, _simple_header, _print_startup_context, _print_root_help, LOGO_ICON, TAGLINE)
 
-
-# ------------------------------------------------------------
-# CLI Process Logging
-# ------------------------------------------------------------
-
-# Maximum log file size (5MB)
-MAX_LOG_SIZE = 5 * 1024 * 1024
-
-
-def _strip_wrapping_quotes(text: Optional[str]) -> Optional[str]:
-    """Remove a single or nested pair of matching wrapping quotes from text.
-
-    Handles common copy/paste cases like "'/Users/me/My Folder'" or '"/path"'.
-    Returns None unchanged and leaves interior quotes untouched.
-    """
-    if text is None:
-        return None
-    s = text.strip()
-    # Remove up to two layers of matching quotes (single or double)
-    for _ in range(2):
-        if len(s) >= 2 and s[0] == s[-1] and s[0] in ("'", '"'):
-            s = s[1:-1].strip()
-        else:
-            break
-    return s
-
-
-def _sanitize_arguments(args: List[str]) -> List[str]:
-    """Sanitize command-line arguments to remove sensitive information."""
-    sanitized = []
-
-    # Patterns for sensitive information
-    sensitive_patterns = [
-        # File paths with potentially sensitive directory names
-        r"(/[Uu]sers?/[^/]+/[Dd]esktop|/[Uu]sers?/[^/]+/[Dd]ocuments)",
-        r"(/home/[^/]+/[Dd]esktop|/home/[^/]+/[Dd]ocuments)",
-        # API tokens and keys
-        r"(--?(?:token|key|password|pass|secret)(?:=|\s+)\S+)",
-        # Email addresses
-        r"\b[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\.[A-Z|a-z]{2,}\b",
-    ]
-
-    for arg in args:
-        sanitized_arg = arg
-        for pattern in sensitive_patterns:
-            if re.search(pattern, arg, re.IGNORECASE):
-                # Replace file paths with just the filename
-                if "/" in arg or "\\" in arg:
-                    path_obj = Path(arg)
-                    sanitized_arg = (
-                        f"[REDACTED]/{path_obj.name}" if path_obj.name else "[REDACTED]"
-                    )
-                else:
-                    # For tokens/keys, show only the parameter name
-                    if "=" in arg:
-                        param_name = arg.split("=")[0]
-                        sanitized_arg = f"{param_name}=[REDACTED]"
-                    else:
-                        sanitized_arg = "[REDACTED]"
-                break
-
-        sanitized.append(sanitized_arg)
-
-    return sanitized
-
-
-def _rotate_log(log_path: Path) -> None:
-    """Rotate log file when it gets too large."""
-    try:
-        # Keep last 5 rotated logs
-        for i in range(4, 0, -1):
-            old_path = log_path.with_suffix(f".{i}.txt")
-            new_path = log_path.with_suffix(f".{i + 1}.txt")
-            if old_path.exists():
-                old_path.rename(new_path)
-
-        # Move current log to .1
-        if log_path.exists():
-            rotated_path = log_path.with_suffix(".1.txt")
-            log_path.rename(rotated_path)
-    except Exception:
-        # If rotation fails, truncate the log
-        try:
-            with log_path.open("w", encoding="utf-8") as f:
-                f.write(
-                    f"[{datetime.now().strftime('%Y-%m-%d %H:%M:%S')}] Log rotated due to size limit\n"
-                )
-        except Exception:
-            pass
-
-
-def _log_cli_execution(args: argparse.Namespace) -> None:
-    """Log CLI execution to workspace process log with security and error handling."""
-    try:
-        # Only log if workspace exists to avoid setup errors
-        workspace_dir = user_config.config_dir
-        if not workspace_dir.exists():
-            return
-
-        log_path = workspace_dir / "process_log.txt"
-        timestamp = datetime.now().strftime("%Y-%m-%d %H:%M:%S")
-
-        # Get original command arguments, excluding the script name
-        original_args = (
-            sys.argv[1:] if hasattr(sys, "argv") and len(sys.argv) > 1 else []
-        )
-
-        # Skip logging if no meaningful command (just bare invocation)
-        if not original_args or not args.command:
-            return
-
-        # Sanitize arguments for security
-        safe_args = _sanitize_arguments(original_args)
-        command_str = f"autocleaneeg-pipeline {' '.join(safe_args)}"
-
-        # Check file size and rotate if necessary
-        if log_path.exists() and log_path.stat().st_size > MAX_LOG_SIZE:
-            _rotate_log(log_path)
-
-        # Atomic write to prevent corruption
-        log_entry = f"[{timestamp}] {command_str}\n"
-
-        # Write to temporary file first, then move
-        temp_path = log_path.with_suffix(".tmp")
-        try:
-            # Read existing content if file exists
-            existing_content = ""
-            if log_path.exists():
-                with log_path.open("r", encoding="utf-8") as f:
-                    existing_content = f.read()
-
-            # Write to temp file
-            with temp_path.open("w", encoding="utf-8") as f:
-                f.write(existing_content + log_entry)
-
-            # Atomic move
-            temp_path.replace(log_path)
-
-        except Exception:
-            # Clean up temp file if it exists
-            if temp_path.exists():
-                temp_path.unlink()
-            raise
-
-    except Exception as e:
-        # Log to stderr but don't break CLI functionality
-        print(f"Warning: Failed to log command execution: {e}", file=sys.stderr)
-
-
-# ------------------------------------------------------------
-# Rich help integration
-# ------------------------------------------------------------
-def _print_startup_context(console) -> None:
-    """Print system info, workspace path, and free disk space (shared for header/help)."""
-    try:
-        from rich.text import Text
-        from rich.align import Align
-        import platform as _platform
-
-        py_ver = f"{sys.version_info.major}.{sys.version_info.minor}.{sys.version_info.micro}"
-        os_name = _platform.system() or "UnknownOS"
-        os_rel = _platform.release() or ""
-        now_str = datetime.now().strftime("%Y-%m-%d %H:%M")
-
-        info = Text()
-        info.append("🐍 Python ", style="muted")
-        info.append(py_ver, style="accent")
-        info.append("  •  ", style="muted")
-        info.append("🖥 ", style="muted")
-        info.append(f"{os_name} {os_rel}".strip(), style="accent")
-        info.append("  •  ", style="muted")
-        info.append("🕒 ", style="muted")
-        info.append(now_str, style="accent")
-        console.print(Align.center(info))
-    except Exception:
-        pass
-
-    # Workspace + disk
-    try:
-        from rich.text import Text as _Text
-        from rich.align import Align as _Align
-
-        workspace_dir = user_config.config_dir
-        try:
-            # Prefer strict validity check (requires saved setup + structure)
-            valid_ws = user_config._is_workspace_valid()  # type: ignore[attr-defined]
-        except Exception:
-            # Fallback to basic existence check
-            valid_ws = workspace_dir.exists() and (workspace_dir / "tasks").exists()
-        home = str(Path.home())
-        display_path = str(workspace_dir)
-        if display_path.startswith(home):
-            display_path = display_path.replace(home, "~", 1)
-
-        ws = _Text()
-        if valid_ws:
-            ws.append("✓ ", style="success")
-            ws.append("Workspace ", style="muted")
-            ws.append(display_path, style="accent")
-            console.print(_Align.center(ws))
-        else:
-            ws.append("⚠ ", style="warning")
-            ws.append("Workspace not configured — ", style="muted")
-            ws.append(display_path, style="accent")
-            console.print(_Align.center(ws))
-            tip = _Text()
-            tip.append("Run ", style="muted")
-            tip.append("autocleaneeg-pipeline workspace", style="accent")
-            tip.append(" to configure.", style="muted")
-            console.print(_Align.center(tip))
-
-        # Always show active task line beneath Workspace (or guard if not set)
-        try:
-            active_task = user_config.get_active_task()
-            at = _Text()
-            at.append("🎯 ", style="muted")
-            at.append("Active task: ", style="muted")
-            if active_task:
-                at.append(str(active_task), style="accent")
-            else:
-                at.append("not set", style="warning")
-            console.print(_Align.center(at))
-        except Exception:
-            pass
-
-        # Show active input (or guard if not set/missing)
-        try:
-            active_source = user_config.get_active_source()
-            src = _Text()
-            if active_source:
-                sp = Path(active_source)
-                display_src = str(sp)
-                home = str(Path.home())
-                if display_src.startswith(home):
-                    display_src = display_src.replace(home, "~", 1)
-                if sp.exists():
-                    if sp.is_file():
-                        src.append("📄 ", style="muted")
-                        src.append("Input file: ", style="muted")
-                    elif sp.is_dir():
-                        src.append("📂 ", style="muted")
-                        src.append("Input folder: ", style="muted")
-                    else:
-                        src.append("📁 ", style="muted")
-                        src.append("Input: ", style="muted")
-                    src.append(display_src, style="accent")
-                else:
-                    src.append("⚠ ", style="warning")
-                    src.append("Input missing — ", style="muted")
-                    src.append(display_src, style="accent")
-            else:
-                src.append("📁 ", style="muted")
-                src.append("Active input: ", style="muted")
-                src.append("not set", style="warning")
-            console.print(_Align.center(src))
-        except Exception:
-            pass
-
-        # Disk free
-        usage_path = (
-            workspace_dir
-            if workspace_dir.exists()
-            else (
-                workspace_dir.parent if workspace_dir.parent.exists() else Path.home()
-            )
-        )
-        du = shutil.disk_usage(str(usage_path))
-        free_gb = du.free / (1024**3)
-        free_line = _Text()
-        free_line.append("💾 ", style="muted")
-        free_line.append("Free space ", style="muted")
-        free_line.append(f"{free_gb:.1f} GB", style="accent")
-        console.print(_Align.center(free_line))
-        console.print()
-    except Exception:
-        pass
-
-
-class RichHelpAction(argparse.Action):
-    """Subparser -h/--help: show styled header + context, then default help."""
-
-    def __call__(self, parser, namespace, values, option_string=None):  # type: ignore[override]
-        console = get_console(
-            namespace if isinstance(namespace, argparse.Namespace) else None
-        )
-        _simple_header(console)
-        _print_startup_context(console)
-        console.print(parser.format_help())
-        sys.exit(0)
-
-
-class RootRichHelpAction(argparse.Action):
-    """Root -h/--help: show styled header + context; supports optional topic like '-h auth'."""
-
-    def __call__(self, parser, namespace, values, option_string=None):  # type: ignore[override]
-        console = get_console(
-            namespace if isinstance(namespace, argparse.Namespace) else None
-        )
-        _simple_header(console)
-        _print_startup_context(console)
-
-        topic = (values or "").strip().lower() if isinstance(values, str) else None
-        _print_root_help(console, topic)
-        sys.exit(0)
-
-
-def _print_root_help(console, topic: Optional[str] = None) -> None:
-    """Print the root help menu with optional topic sections, in a clean minimalist layout."""
-    from rich.table import Table as _Table
-
-    # Compact usage line for quick orientation
-    console.print(
-        "[muted]Usage:[/muted] [accent]autocleaneeg-pipeline <command> [options][/accent]"
-    )
-    console.print()
-
-    if topic in {"auth", "authentication"}:
-        console.print("[header]Auth Commands[/header]")
-        tbl = _Table(show_header=False, box=None, padding=(0, 1))
-        tbl.add_column("Command", style="accent", no_wrap=True)
-        tbl.add_column("Description", style="muted")
-
-        rows = [
-            ("🔐 auth login", "Login to Auth0 (compliance mode)"),
-            ("🔓 auth logout", "Logout and clear tokens"),
-            ("👤 auth whoami", "Show authenticated user"),
-            ("🩺 auth diagnostics", "Diagnose Auth0 configuration/connectivity"),
-            ("⚙️ auth setup", "Enable Part-11 compliance (permanent)"),
-            ("🟢 auth enable", "Enable compliance mode (non-permanent)"),
-            ("🔴 auth disable", "Disable compliance mode (if permitted)"),
-        ]
-        for c, d in rows:
-            tbl.add_row(c, d)
-        console.print(tbl)
-        console.print(
-            "[muted]Docs:[/muted] [accent]https://docs.autocleaneeg.org[/accent]"
-        )
-        console.print()
-        return
-
-    if topic in {"task", "tasks"}:
-        console.print("[header]Task Commands[/header]")
-        tbl = _Table(show_header=False, box=None, padding=(0, 1))
-        tbl.add_column("Command", style="accent", no_wrap=True)
-        tbl.add_column("Description", style="muted")
-        rows = [
-            ("📜 task list", "List available tasks (same as 'list-tasks')"),
-            ("📂 task explore", "Open the workspace tasks folder"),
-            ("✏️  task edit [name|path]", "Edit task (omit uses active)"),
-            ("📥 task import <path>", "Copy a task file into workspace"),
-            ("📄 task copy [name|path]", "Copy task (omit uses active)"),
-            ("🗑  task delete [name|path]", "Delete task (omit uses active)"),
-            ("🎯 task set [name]", "Set active task (interactive if omitted)"),
-            ("🧹 task unset", "Clear the active task"),
-            ("👁️  task show", "Show the current active task"),
-        ]
-        for c, d in rows:
-            tbl.add_row(c, d)
-        console.print(tbl)
-        console.print()
-        console.print(
-            "[muted]Docs:[/muted] [accent]https://docs.autocleaneeg.org[/accent]"
-        )
-        console.print()
-        return
-
-    if topic in {"input", "inputs"}:
-        console.print("[header]Input Commands[/header]")
-        tbl = _Table(show_header=False, box=None, padding=(0, 1))
-        tbl.add_column("Command", style="accent", no_wrap=True)
-        tbl.add_column("Description", style="muted")
-        rows = [
-            (
-                "📁 input set [path]",
-                "Set active input path (file or directory; interactive if omitted)",
-            ),
-            ("🧹 input unset", "Clear the active input path"),
-            ("👁️  input show", "Show the current active input path"),
-        ]
-        for c, d in rows:
-            tbl.add_row(c, d)
-        console.print(tbl)
-        console.print()
-        console.print(
-            "[muted]Docs:[/muted] [accent]https://docs.autocleaneeg.org[/accent]"
-        )
-        console.print()
-        return
-
-    if topic in {"source", "sources"}:
-        console.print(
-            "[header]Source Commands[/header] [warning](deprecated — use 'input')[/warning]"
-        )
-        tbl = _Table(show_header=False, box=None, padding=(0, 1))
-        tbl.add_column("Command", style="accent", no_wrap=True)
-        tbl.add_column("Description", style="muted")
-        rows = [
-            ("📁 source set [path]", "Alias of 'input set' (interactive if omitted)"),
-            ("🧹 source unset", "Alias of 'input unset'"),
-            ("👁️  source show", "Alias of 'input show'"),
-        ]
-        for c, d in rows:
-            tbl.add_row(c, d)
-        console.print(tbl)
-        console.print()
-        console.print(
-            "[muted]Docs:[/muted] [accent]https://docs.autocleaneeg.org[/accent]"
-        )
-        console.print()
-        return
-
-    if topic in {"workspace", "setup"}:
-        console.print("[header]Workspace[/header]")
-        tbl = _Table(show_header=False, box=None, padding=(0, 1))
-        tbl.add_column("Command", style="accent", no_wrap=True)
-        tbl.add_column("Description", style="muted")
-        rows = [
-            ("🗂  workspace", "Configure workspace folder (wizard)"),
-            ("👀 workspace show", "Show current workspace path/status"),
-            ("📂 workspace explore", "Open the workspace folder"),
-            ("📏 workspace size", "Show total workspace size"),
-            ("📌 workspace set <path>", "Change the workspace folder"),
-            ("❎ workspace unset", "Unassign current workspace (clear config)"),
-            ("📁 workspace cd [--spawn]", "Print path for cd, or spawn subshell"),
-            ("🏠 workspace default", "Set recommended default location"),
-            ("—", "—"),
-            ("🔐 auth setup|enable|disable", "Compliance controls (Auth0)"),
-        ]
-        for c, d in rows:
-            tbl.add_row(c, d)
-        console.print(tbl)
-        console.print()
-        console.print(
-            "[muted]Docs:[/muted] [accent]https://docs.autocleaneeg.org[/accent]"
-        )
-        console.print()
-        return
-
-    console.print("[header]Commands[/header]")
-    tbl = _Table(show_header=False, box=None, padding=(0, 1))
-    tbl.add_column("Command", style="accent", no_wrap=True)
-    tbl.add_column("Description", style="muted")
-
-    rows = [
-        ("❓ help", "Show help and topics (alias for -h/--help)"),
-        ("🗂\u00a0 workspace", "Configure workspace folder"),
-        ("👁\u00a0 view", "View EEG file (MNE-QT)"),
-        ("🗂\u00a0 task", "Manage tasks (list, explore)"),
-        ("📁\u00a0 input", "Manage active input path"),
-        ("▶\u00a0 process", "Process EEG data"),
-        ("📝 review", "Start review GUI"),
-        ("🔐 auth", "Authentication & Part-11 commands"),
-    ]
-    for c, d in rows:
-        tbl.add_row(c, d)
-    console.print(tbl)
-    console.print()
-    console.print("[muted]Docs:[/muted] [accent]https://docs.autocleaneeg.org[/accent]")
-    console.print()
-
-
-def attach_rich_help(p: argparse.ArgumentParser, *, root: bool = False) -> None:
-    # Replace default help with our rich-aware action
-    if any(a.option_strings == ["-h"] for a in p._actions):  # remove default
-        for a in list(p._actions):
-            if a.option_strings == ["-h"] or a.option_strings == ["-h", "--help"]:
-                p._actions.remove(a)
-                break
-    action = RootRichHelpAction if root else RichHelpAction
-    nargs = "?" if root else 0
-    p.add_argument(
-        "-h",
-        "--help",
-        action=action,
-        nargs=nargs,
-        help='Show help (use "-h auth" for authentication help)',
-    )
-
-
-# Simple branding constants
-PRODUCT_NAME = "AutoClean EEG"
-TAGLINE = "Automated EEG Processing Software"
-LOGO_ICON = "🧠"
-DIVIDER = "═════════════════════════════════════════════════"
 
 # Try to import database functions (used conditionally in login)
 try:
@@ -2435,33 +1953,6 @@ def cmd_workspace_cd(args) -> int:
         return 1
 
 
-def _simple_header(
-    console, title: Optional[str] = None, subtitle: Optional[str] = None
-):
-    """Simple, consistent header for setup."""
-    from rich.align import Align
-    from rich.panel import Panel
-    from rich.text import Text
-
-    console.print()
-
-    # Create branding content (no borders; app name with version)
-    branding_text = Text()
-    branding_text.append(
-        f"{LOGO_ICON} AutocleanEEG Pipeline ({__version__})", style="brand"
-    )
-    branding_text.append(f"\n{TAGLINE}", style="accent")
-
-    # Print centered branding (no borders)
-    console.print(Align.center(branding_text))
-    console.print()
-    if title:
-        console.print(f"[title]{title}[/title]")
-    if subtitle:
-        console.print(f"[subtitle]{subtitle}[/subtitle]")
-    console.print()
-
-
 def _run_interactive_setup() -> int:
     """Run interactive setup wizard with arrow key navigation."""
 
@@ -3992,7 +3483,7 @@ def cmd_source_set(args) -> int:
     """Set the active input path (stored internally as 'source')."""
     try:
         source_path = getattr(args, "source_path", None)
-        source_path = _strip_wrapping_quotes(source_path)
+        source_path = strip_wrapping_quotes(source_path)
 
         # If path provided directly, use it
         if source_path:
@@ -4010,7 +3501,7 @@ def cmd_source_set(args) -> int:
 
         # Otherwise, use interactive selection
         selected_source = user_config.select_active_source_interactive()
-        selected_source = _strip_wrapping_quotes(selected_source)
+        selected_source = strip_wrapping_quotes(selected_source)
         if selected_source is None:
             message("info", "Canceled")
             return 0
@@ -5245,7 +4736,7 @@ def main(argv: Optional[list] = None) -> int:
     # ------------------------------------------------------------------
     # Log CLI execution for process tracking
     # ------------------------------------------------------------------
-    _log_cli_execution(args)
+    log_cli_execution(args)
 
     # ------------------------------------------------------------------
     # Always inform the user where the AutoClean workspace is (or will be)
